init commit I should have committed at the start of project, but SourceTree was being a pain by messing up authentication to GitHub and I didn't want to fix it right now.
Some of this may supposed to be in the template project, but I was trying to work on them both at the same time instead of focussing on just one

So now I'm just committing this almost completed backend in one big commit. I'll continue to remove the rest of the template project code in future commits, but for now I'm just pushing the whole thing so that I've got a "blank" slate for commits for the rest of this project.

TODO for the next project:
* fix SourceTree auth problems, and then commit things in smaller chunks
* only work on one project at a time (or use branches to separate out the bits I'm working on)